opal-ci: Update qemu command line to test booting up
Starting with qemu 7.2, there's no longer any PCI root port configured
when the '-nodefaults' option is used. Since that option is used and
we include some PCI devices when we test booting with qemu, qemu 7.2
is throwing an error as it doesn't have a bus to attach the devices
to. It's currently breaking the CI on Fedora rawhide, where qemu has
been updated to version 7.2.

Support for powernv in qemu has improved since the qemu-boot-check
test was introduced and there's no good reason to run with
'-nodefaults' any more. So this patch removes it. Which allows to also
remove a few extra devices from the command line since the defaults
are sane. And we can still attach PCI devices.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>