hw/chiptod: allow qemu to use the chiptod
There is a QEMU model for ChipTOD (not yet upstream) that provides
enough to get through skiboot's sync sequence.

So run the chiptod init when running under QEMU, if it is being
provided in the device tree. If not, don't crash because the QEMU
patch has not been merged yet...

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>