core/pldm: Add PLDM responder support
PLDM defines data representations and commands that abstract the platform
management hardware.

A PLDM Terminus (or responder) is defined as the point of communication
termination for PLDM messages and the PLDM functions associated with those
messages.
A PLDM terminus is not required to process more than one request at a time
(that is, it can be "single threaded" and does not have to accept and act
on new requests until it has finished responding to any previous request).

Some PLDM control and discovery requests (PLDM_TYPE = PLDM_BASE) are
mandatory a PLDM terminus has to answer.

These following mandatory PLDM command codes for PLDM messaging control
and discovery will be defined in next patches.
    GetTID          0x02
    GetPLDMVersion  0x03
    GetPLDMTypes    0x04
    GetPLDMCommands 0x05

Reviewed-by: Abhishek Singh Tomar <abhishek@linux.ibm.com>
Signed-off-by: Christophe Lombard <clombard@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>