core/pldm: PLDM for Platform Monitoring and Control Specification
This specification defines the functions and data structures used for
discovering, describing, initializing, and accessing sensors and effecters
within the management controllers and management devices of a platform
management subsystem using PLDM messaging.

A PDR (Platform Descriptor Record) is a set of data that is used to
provide semantic information about sensors, effecters, monitored or
controller entities, and functions and services within a PLDM
implementation.
PDRs are mostly used to support PLDM monitoring and control and platform
events.

The PDRs for a PLDM subsystem are collected into a single, central PDR
Repository. A central repository provides a single place from which PDR
information can be retrieved.

The GetPDR command is used to retrieve individual PDRs from a PDR
Repository. The record is identified by the PDR recordHandle value that is
passed in the request.

The patch dump all the PDRs within a PDR Repository.

Reviewed-by: Abhishek Singh Tomar <abhishek@linux.ibm.com>
Signed-off-by: Christophe Lombard <clombard@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>