core/pldm: Encode state effecter request
PLDM effecters provide a general mechanism for controlling or configuring a
state or numeric setting of an entity.
The entity state and numeric setting values are written into an effecter.

PLDM commands are specified for writing the state or numeric setting to an
effecter. Effecters are identified by and accessed using an EffecterID that
is unique for each effecter within a given terminus.

PLDM State Effecters provide a regular command structure for setting state
information in order to change the state of an entity.

The SetStateEffecterStates command is used to set the state of one or more
effecters within a PLDM State Effecter.

Reviewed-by: Abhishek Singh Tomar <abhishek@linux.ibm.com>
Signed-off-by: Christophe Lombard <clombard@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>