core/pldm: Fix pdr handle to add first pdr request
As per the specification:
To retrieve the first PDR record, use the
get_pdr_req function with handle 0.

On the BMC side, the first PDR is sent in
response, along with the next_record_hndl which
can be used to access consecutive PDR records.

However, it's important to note that the first
PDR may not necessarily have a handle of 1.

In the current scenario, providing a record_hndl
value of 0 to pldm_pdr_add() will always result
in the addition of a record to the repository
with a PDR handle of 1.

In current fix record handle is extracted from
pdr record data.

Signed-off-by: Abhishek Singh Tomar <abhishek@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>