core/cpu: Fix invalid SMT indexes
Code using cpu_thread_count is dangerous because that is the maximum
number of threads that a CPU type supports, not the actual number of
threads. For real hardware that hardly matters, but QEMU can run a
single thread Power10, for example. This causes some code (e.g.,
xive_init_cpu_properties) to access beyond the end of allocated
arrays.

Fix this by making cpu_thread_count the actual number of threads
discovered via dt.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>