hw/bt: Work around QEMU lost interrupt bug
QEMU has a bug where it loses the BT interrupt somewhere between BT
and XIVE when the OS boots.

For now, add a workaround QEMU quirk in the poller to try to kick
things along again.

SBE suffers the same problem but it has a poller that kicks the SBE and
gets it going again. Suspect the PSI interrupts may not be re-presented
after the OS re-initialises XIVE. This issue does not seem to appear on
real hardware.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>