hw/sbe-p9: P10 additions
P10 has a lower minimum timeout threshold than P9 (100usecs).

Some P10 SBE timers run about 6.7% slow, which must be a hardware or
firmware issue. Use the SBE timer health checking code to detect this
and compensate for it. Speeding up timers as a rule is dangerous because
early-expiry is a bug, howerver the core timer code checks expiry against
the CPU's timebase when running timers, and with the previous changes it
will schedule a new SBE timer for the remaining delay. So if this
adjustment speeds things up slightly too much, it won't cause bugs.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>