hw/sbe: Add SBE quirk for mambo and awan
There appears to be no device-tree test for the P9 SBE presence like
there is for P8. The P9 device tree test looks for the "primary"
property, but this doesn't really test SBE presence because all chips
have an SBE. It just happens to work because mambo must not add that
property.

So add a platform quirk, and mark mambo and awan as not having SBE.

This is needed for a later change that runs a health check on every
SBE in the system.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
[arbab: Add #include <chip.h>]
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>