external/ffspart: Avoid makefile race condition
In ffspart we assign this make variable:

  FFSPART_VERSION ?= $(shell ./make_version.sh $(EXE))

However, ./make_version.sh is actually a make target, and whether it
exists or not at the time of this assignment is by chance, depending on
how the make concurrency works out.

In practice, this manifests as intermittent CI build failure:

     [ RUN-TEST ]   check-ffspart
     [ RUN-TEST ]   check-pflash
     [ HOSTCC ]  hdata/test/stubs.c
  make[1]: ./make_version.sh: No such file or directory
     [ HOSTCC ]  hw/test/phys-map-test.c
     [ HOSTCC ]  hw/test/run-port80h.c
     [ HOSTCC ]  hw/ipmi/test/run-fru.c
     [ HOSTCC ]  libc/test/run-time.c
     [ HOSTCC ]  libc/test/run-snprintf.c
     [ HOSTCC ]  libc/test/run-snprintf-test.c
     [ HOSTCC ]  libc/test/run-memops.c
  make[1]: *** [Makefile:13: check] Error 1
  make[1]: Entering directory '/build/external/ffspart'

The rule for make_version.sh is just a symlink:

  make_version.sh:
  	$(Q_LN)ln -sf ../../make_version.sh

To avoid the race, call make_version.sh from its actual location instead
of relying on the link to be created. This seems like the motivation for
what was done for gard in commit 8ab0caf26de9 ("external/gard: Fix make
dist target")

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>