extract-gcov: Fix tag lengths
Starting with GCC 12, gcov cannot parse the .gcda files we generate:

  $ powerpc64le-linux-gcov-dump platforms/qemu/qemu.gcda
  platforms/qemu/qemu.gcda:data:magic `adcg':version `B23*' (swapped endianness)
  platforms/qemu/qemu.gcda:stamp 1742688024
  platforms/qemu/qemu.gcda:checksum 2623079854
  platforms/qemu/qemu.gcda:  01000000:   3:FUNCTION ident=1191288390, lineno_checksum=0xdb12f55c, cfg_checksum=0xf9e50e8f
  platforms/qemu/qemu.gcda:tag `46db12f5' is incorrectly nested
  platforms/qemu/qemu.gcda:    46db12f5:1559880974:UNKNOWN

This is due to gcc commit 23eb66d1d46a ("gcov: Use system IO
buffering"), where the length field of tags in the file changed to
represent total bytes, not a count of words.

Change what we write accordingly.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>