hw/lpc: Fix firmware memory space access for devices > 0
The LPC firmware memory space is 256MB in size, and it may select
up to 16 devices with the IDSEL field.

OPB addresses FW space as a 32-bit value with the top 4 bit selecting
the device and the bottom 28 addressing the FW memory space. Therefore
the top bits should ignored when calculating the offset into the FW
window.

Fix this by allowing lpc_opb_prepare() to adjust the address directly
and correctly mask it. Now there's no need to return opb_base to the
caller either, fold that in at the same time.

This bug could be observed with QEMU's PNOR implementation that placed
some of the PNOR in device 1, though that has been changed in QEMU 10.0.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>