core/timers: Fix running-timer delay
When timer run code encounters an alreay-running timer, it has to
stop processing and run them later. In the case of poll timers the
SBE timer is scheduled for a minimum-delay, and for delay timers
nothing is done.

This looks backwards: poll timers do not get called from the
SBE interrupt so that delay is pointless, whereas it is helpful
for delay timers to ensure they're processed again soon.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>