hw/xive: Warn if pushing HW context to already enabled CAM
The push-context operation is not defined if a context is already
valid, it should only be performed if the CAM is pulled. Add a check
to ensure the TIMA reset was performed properly before pushing a
context.

QEMU does not yet model the reset via PTER toggle correctly, so this
causes some noise in boot.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>