astbmc/ipmi: Improve Set Enables handling
The IPMI Set Enables command is supposed to do a read-modify-write
operation to change bits if it is not coded specifically for the
system. Since the various BMCs supported by astbmc platform code
(e.g., QEMU and OpenBMC) are a bit different and subject to change,
it's safer to set bits with RMW.

Then bits should be set one by one to help isolate failures. And
the Set Enables command is changed to run synchronously so that
host/BMC behaviour is a bit more deterministic when setting up IPMI.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>