uart: Fix uninitialised comparison
According to doc/opal-api/opal-console-read-write-1-2.rst, the length
argument of OPAL_CONSOLE_WRITE_BUFFER_SPACE is only used to return a
value. Indeed, the API is called twice in the kernel code, and __length
remains uninitialized in both cases. This can lead to a hang/softlock
issue in older hardware.

Eliminate the problematic comparison which uses the uninitialized value.

Fixes: 6bf21350da32 ("uart: Drop console write data if BMC becomes unresponsive")
Signed-off-by: Abhishek Singh Tomar <abhishek@linux.ibm.com>
Reviewed-by: Aditya Gupta <adityag@linux.ibm.com>
Signed-off-by: Reza Arbab <arbab@linux.ibm.com>